Create Patch_DiscordManager.cs
DarkCactus and hirnukuono noticed that any harmony patch, prefix bool or postfix void, even empty, will fix the "return all players to lobby" -button problem, and Discord Rich Presence seems to work fine even on modded rundowns with id 1.

Had a nice configurable conditional patch working and was trying to configure it as disabled / false, but the button still appeared every time. Led to the above discovery and a much simpler fix to the problem.